+++ packages/suite-build/utils/constants.ts
@@ -1,7 +1,8 @@
-export const PROJECTS = ['web', 'desktop'] as const;
+export const PROJECTS = ['web', 'desktop', 'landing'] as const;
 export type Project = typeof PROJECTS[number];
 
 export const DEV_PORTS: { [key in Project]: number } = {
     web: 8000,
     desktop: 8000,
+    landing: 3000,
 };
