--- packages/suite-web-landing/pages/_app.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import NextApp from 'next/app';
-import { ThemeProvider } from 'styled-components';
-import { THEME } from '@trezor/components';
-
-class App extends NextApp {
-    render() {
-        const { Component, pageProps } = this.props;
-        return (
-            <>
-                <ThemeProvider theme={THEME.light}>
-                    <Component {...pageProps} />
-                </ThemeProvider>
-            </>
-        );
-    }
-}
-
-export default App;
