+++ packages/suite-web-landing/src/global.d.ts
@@ -1,0 +1,3 @@
+/* eslint-disable @typescript-eslint/triple-slash-reference */
+/// <reference path="../../suite/global.d.ts" />
+/// <reference path="../../suite/styled.d.ts" />
