+++ packages/suite-web-landing/src/static/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en" translate="no">
+    <head>
+        <meta name="google" content="notranslate" />
+        <meta charset="utf-8">
+        <title>Trezor Suite</title>
+
+        <link
+            media="all"
+            rel="stylesheet"
+            href="<%= assetPrefix %>/static/fonts/fonts.css"
+        />
+        <link rel="icon" href="<%= assetPrefix %>/static/images/icons/favicon/favicon.png" />
+        <link
+            rel="apple-touch-icon"
+            href="<%= assetPrefix %>/static/images/icons/favicon/favicon.png"
+        />
+        <script src="<%= assetPrefix %>/static/browser-detection/index.js"></script>
+    </head>
+    <body>
+        <noscript>This website requires JavaScript to run.</noscript>
+    </body>
+</html>
